requirements.txt
urllib3==1.26.16
requests==2.28.2
Pillow==10.0.1
beautifulsoup4==4.12.2
selenium-wire==5.1.0